work: better fix

+++ solid/utility/test/test_workpool_basic.cpp
@@ -19,14 +19,14 @@ int test_workpool_basic(int /*argc*/, char* /*argv*/ [])
     using WorkPoolT  = WorkPool<size_t>;
     using AtomicPWPT = std::atomic<WorkPoolT*>;
 
-    const int             wait_seconds = 500;
-    const int             loop_cnt     = 10;
-    const size_t          cnt{5000000};
-    const size_t          v = (((cnt - 1) * cnt)) / 2;
-    std::atomic<uint64_t> val{0};
-    promise<void>         prom;
-    AtomicPWPT            pwp{nullptr};
-    thread                wait_thread(
+    const int           wait_seconds = 500;
+    const int           loop_cnt     = 10;
+    const size_t        cnt{5000000};
+    const size_t        v = (((cnt - 1) * cnt)) / 2;
+    std::atomic<size_t> val{0};
+    promise<void>       prom;
+    AtomicPWPT          pwp{nullptr};
+    thread              wait_thread(
         [](promise<void>& _rprom, AtomicPWPT& _rpwp, const int _wait_time_seconds) {
             if (_rprom.get_future().wait_for(chrono::seconds(_wait_time_seconds)) != future_status::ready) {
                 if (_rpwp != nullptr) {
